refactoring code and test; add interface AlgorithmRecursive/change on AlgorithmRecursiveImpl
