+++ src/algorithm/AlgorithmRecursive.java
@@ -5,91 +5,17 @@
  */
 package algorithm;
 
-import algorithm.config.AlgorithmRecursiveConfiguration;
-import algorithm.exception.IterationLimitHasBeenExceededException;
-import algorithm.modifier.GenInstance;
-import algorithm.modifier.ModifierResultRecursive;
 import algorithm.modifier.ResultRecursive;
-import java.util.ArrayList;
-import java.util.List;
 import algorithm.node.NodeAlgorithm;
 
 /**
  *
- * @author kamil.jarmusik
+ * @author Kamil-Tomasz
  * @param <D>
  * @param <R>
  */
-public final class AlgorithmRecursive<D, R> {
-    
-    private final AlgorithmRecursiveConfiguration<D> config;
-    private final ModifierResultRecursive<D, R> modifier;
-    private NodeAlgorithm<D> start;
-
-    AlgorithmRecursive(NodeAlgorithm<D> start, AlgorithmRecursiveConfiguration<D> config, ModifierResultRecursive<D, R> modifier) {
-        this.config = config;
-        this.modifier = modifier;
-        this.start = start;
-    }
-
-    public static <D, R> AlgorithmRecursiveImmutableResultBuilder<D, R> immutableResult(NodeAlgorithm<D> start, R result) {
-        return new AlgorithmRecursiveImmutableResultBuilder<>(start, result);
-    }
-    
-    public static <D, R> AlgorithmRecursiveMutableResultBuilder<D, R> mutableResult(NodeAlgorithm<D> start, GenInstance<R> instanceGenerator) {
-        return new AlgorithmRecursiveMutableResultBuilder<>(start, instanceGenerator);
-    }
-    
-    public static <D> AlgorithmRecursiveMutableResultBuilder<D, List<D>> aggregateResult(NodeAlgorithm<D> start) {
-        return new AlgorithmRecursiveMutableResultBuilder<>(start, () -> new ArrayList<>());
-    }
-    
-    public void changeStartAndRun(NodeAlgorithm<D> start) {
-        this.start = start;
-        run();
-    }
-    
-    public ResultRecursive<R> result() {
-        return modifier.snapshot();
-    }
-    
-    public void run() {
-        modifier.reset();
-        modifier.incrementIteration();
-        if(config.isToExecute(start))
-            modifier.execute(start.data());
-        if(isEnd(start))
-            return;
-        step(start);
-    }
-
-    private boolean isEnd(NodeAlgorithm<D> node) {
-        return isLeaf(node) || config.isFinish(node);
-    }
-    
-    private static <T> boolean isLeaf(NodeAlgorithm<T> node) {
-        List<NodeAlgorithm<T>> nodes = node.nodes();
-        return nodes == null || nodes.isEmpty();
-    }
-    
-    private void step(NodeAlgorithm<D> start) {
-        modifier.incrementIteration();
-        ifHasBeenExceededThenThrowException();
-        start.nodes().stream().map(node -> {
-            if(config.isToExecute(node))
-                modifier.execute(node.data());
-            return node;
-        }).filter(node -> !isEnd(node)).forEachOrdered(node -> {
-            step(node);
-        }); 
-    }
-
-    private void ifHasBeenExceededThenThrowException() {
-        if(hasBeenExceeded())
-            throw new IterationLimitHasBeenExceededException(config.getLimitNumberIterations(), modifier.snapshot().getNumberIteration());
-    }
-
-    private boolean hasBeenExceeded() {
-        return config.getLimitNumberIterations() > 0 && modifier.hasBeenExceeded(config.getLimitNumberIterations());
-    }
+public interface AlgorithmRecursive<D, R> {
+    void run();
+    void changeStartAndRun(NodeAlgorithm<D> start);
+    ResultRecursive<R> result();
 }
