+++ src/impl/AlgorithmRecursiveImplFactory.java
@@ -6,6 +6,7 @@
 package impl;
 
 import algorithm.AlgorithmRecursive;
+import algorithm.AlgorithmRecursiveFactory;
 import impl.file.NodeFile;
 import impl.number.NodeInteger;
 import impl.file.ext.Extension;
@@ -19,10 +20,10 @@ import impl.file.util.FileUtil;
  *
  * @author Kamil-Tomasz
  */
-public final class AlgorithmRecursiveFactory {
+public final class AlgorithmRecursiveImplFactory {
     
     public static AlgorithmRecursive<File, List<File>> aggregateFiles(File dir) {
-        return AlgorithmRecursive.aggregateResult(new NodeFile(dir))
+        return AlgorithmRecursiveFactory.aggregateResult(new NodeFile(dir))
                         .finishIf(node -> false)
                         .executeIf(node -> !FileUtil.isDirectory(node.data()))
                         .toExecute((file,result) -> result.add(file))
@@ -30,7 +31,7 @@ public final class AlgorithmRecursiveFactory {
     }
     
     public static AlgorithmRecursive<File, List<File>> aggregateFilesByName(File dir, String name) {
-        return AlgorithmRecursive.aggregateResult(new NodeFile(dir))
+        return AlgorithmRecursiveFactory.aggregateResult(new NodeFile(dir))
                         .finishIf(node -> false)
                         .executeIf(node -> !FileUtil.isDirectory(node.data()) 
                                 && node.data().getName().contains(name))
@@ -39,7 +40,7 @@ public final class AlgorithmRecursiveFactory {
     }
     
     public static AlgorithmRecursive<File, List<File>> aggregateFilesByExtension(File dir, Extension extension) {
-        return AlgorithmRecursive.aggregateResult(new NodeFile(dir))
+        return AlgorithmRecursiveFactory.aggregateResult(new NodeFile(dir))
                         .finishIf(node -> false)
                         .executeIf(node -> !FileUtil.isDirectory(node.data()) 
                                 && Extension.ext(node.data()) == extension)
@@ -48,7 +49,7 @@ public final class AlgorithmRecursiveFactory {
     }
     
     public static AlgorithmRecursive<File, List<File>> aggregateFilesByExtensions(File dir, Set<Extension> extension) {
-        return AlgorithmRecursive.aggregateResult(new NodeFile(dir))
+        return AlgorithmRecursiveFactory.aggregateResult(new NodeFile(dir))
                         .finishIf(node -> false)
                         .executeIf(node -> !FileUtil.isDirectory(node.data()) 
                                 && extension.contains(Extension.ext(node.data())))
@@ -57,7 +58,7 @@ public final class AlgorithmRecursiveFactory {
     }
     
     public static AlgorithmRecursive<Integer, BigDecimal> factorial(int arg) {
-        return AlgorithmRecursive.immutableResult(new NodeInteger(arg), BigDecimal.ONE)
+        return AlgorithmRecursiveFactory.immutableResult(new NodeInteger(arg), BigDecimal.ONE)
                         .finishIf(node -> node.data() < 2)
                         .executeIf(node -> true)
                         .limitNumberIterations(700)
